requirements.txt
flask==2.0.1
Werkzeug==2.0.1
requests==2.31.0
python-dotenv==1.0.0